text_llm_pipeline/requirements.txt
guru4elephant-jsonflow>=0.1.0
requests>=2.28.0
tqdm>=4.64.0
argparse>=1.4.0
pyyaml>=6.0 